split the dashboard into components so we can parallelize work better
