+++ products/ai_visibility/frontend/components/TopCitedSourcesList.tsx
@@ -1,0 +1,29 @@
+import { Link } from 'lib/lemon-ui/Link'
+
+import { TopCitedSource } from '../types'
+
+export function TopCitedSourcesList({ sources }: { sources: TopCitedSource[] }): JSX.Element {
+    return (
+        <div className="border rounded-lg p-4 bg-bg-light">
+            <div className="flex items-center justify-between mb-4">
+                <h3 className="text-sm font-semibold">Top cited sources</h3>
+                <Link to="#" className="text-xs text-primary">
+                    View all
+                </Link>
+            </div>
+            <div className="space-y-3">
+                {sources.slice(0, 6).map((source) => (
+                    <div key={source.domain} className="flex items-center justify-between">
+                        <div className="flex items-center gap-2">
+                            <div className="w-5 h-5 rounded bg-border flex items-center justify-center text-xs">🌐</div>
+                            <span className="text-sm">{source.domain}</span>
+                        </div>
+                        <span className="text-sm">
+                            <strong>{source.responseCount}</strong> <span className="text-muted">responses</span>
+                        </span>
+                    </div>
+                ))}
+            </div>
+        </div>
+    )
+}
